@@ -1,6 +1,6 @@
 module github.com/silenceper/wechat/v2
 
-go 1.15
+go 1.16
 
 require (
 	github.com/bradfitz/gomemcache v0.0.0-20220106215444-fb4bf637b56d
